Enhance dualboot support and update configuration handling
- Added dualboot option in `customize_config` to allow users to specify dualboot settings during configuration.
- Updated `print_config_data` to display the dualboot status.
- Modified command-line arguments in `main.py` to include a flag for enabling dualboot mode.
- Refactored `format_partitions` and `install_grub` functions to handle dualboot scenarios, ensuring proper installation of GRUB and partition formatting based on user input.
- Updated minimal and standart configuration files to include a default dualboot setting.

These changes improve the installation process by providing users with the ability to configure dualboot options seamlessly.

+++ modules/disk.py
@@ -51,12 +51,13 @@ def markup_disk(disk, partitions_size, partitions_format):
     Print.success("Partitioning complete")
     print()
 
-def format_partitions(partitions, partitions_format):
+def format_partitions(partitions, partitions_format, dualboot):
     Print.info("Formatting disks...")
-    if partitions_format['boot'].lower() == "fat32":
-        run_cmd(f"mkfs.fat -F 32 {partitions['boot']}")
-    else:
-        run_cmd(f"mkfs.{partitions_format['boot']} {partitions['boot']}")
+    if not dualboot:
+        if partitions_format['boot'].lower() == "fat32":
+            run_cmd(f"mkfs.fat -F 32 {partitions['boot']}")
+        else:
+            run_cmd(f"mkfs.{partitions_format['boot']} {partitions['boot']}")
     run_cmd(f"mkswap {partitions['swap']}")
     run_cmd(f"swapon {partitions['swap']}")
     if partitions_format['home'].lower() == "btrfs":
